Move the Debris object back into objectDetection
