+++ src/test/java/fpga/ExtractTest.java
@@ -4,7 +4,6 @@ import static fpga.Testing.assertEquals;
 import static fpga.Testing.assertTrue;
 
 import fpga.objectdetection.Debris;
-
 import java.io.IOException;
 import java.util.List;
 
@@ -21,12 +20,12 @@ public class ExtractTest extends Extract
     boolean[][] image = new boolean[300][300];
     image[50][45] = true; //Single pixel asteroid at 50x45
 
-    List<Extract.Debris> debrisList = Extract.extract(image);
+    List<Debris> debrisList = Extract.extract(image);
 
 
     assertEquals(new Integer(1), debrisList.size());
 
-    Extract.Debris debris = debrisList.get(0);
+    Debris debris = debrisList.get(0);
 
     assertEquals(new Integer(50), debris.centerX);
     assertEquals(new Integer(45), debris.centerY);
@@ -39,12 +38,12 @@ public class ExtractTest extends Extract
     image[250][200] = true; //Single pixel asteroid at 50x45
 
 
-    List<Extract.Debris> debrisList = Extract.extract(image);
+    List<Debris> debrisList = Extract.extract(image);
 
     assertEquals(new Integer(2), debrisList.size());
 
-    assertTrue(debrisList.contains(new Extract.Debris(50,45,1)));
-    assertTrue(debrisList.contains(new Extract.Debris(250,200,1)));
+    assertTrue(debrisList.contains(new Debris(50,45,1)));
+    assertTrue(debrisList.contains(new Debris(250,200,1)));
   }
 
   public void cornerCaseDebrisTest() throws Exception {
@@ -56,11 +55,11 @@ public class ExtractTest extends Extract
     }
 
 
-    List<Extract.Debris> debrisList = Extract.extract(image);
+    List<Debris> debrisList = Extract.extract(image);
 
     assertEquals(new Integer(1), debrisList.size());
 
-    Extract.Debris debris = debrisList.get(0);
+    Debris debris = debrisList.get(0);
 
     assertEquals(new Integer(25), debris.centerX);
     assertEquals(new Integer(25), debris.centerY);
